requirements.txt
django-jsonfield
djangorestframework
django
djangorestframework-simplejwt